keep clean main.go, multiplexing init func
