+++ business/config/config.go
@@ -1,0 +1,226 @@
+package config
+
+import (
+	"context"
+	"expvar"
+	"fmt"
+	"net/http"
+	"os"
+	"os/signal"
+	"syscall"
+
+	"github.com/ardanlabs/conf"
+	"github.com/ardanlabs/service/app/sales-api/handlers"
+	"github.com/ardanlabs/service/business/sys/auth"
+	"github.com/ardanlabs/service/business/sys/metrics"
+	"github.com/ardanlabs/service/foundation/database"
+	"github.com/ardanlabs/service/foundation/keystore"
+	"github.com/jmoiron/sqlx"
+	"github.com/pkg/errors"
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/exporters/zipkin"
+	"go.opentelemetry.io/otel/sdk/resource"
+	"go.opentelemetry.io/otel/sdk/trace"
+	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
+	"go.uber.org/zap"
+)
+
+func Configuration(log *zap.SugaredLogger, build string) (Config, error) {
+	cfg := Config{
+		Version: conf.Version{
+			SVN:  build,
+			Desc: "copyright information here",
+		},
+	}
+	if err := conf.Parse(os.Args[1:], "SALES", &cfg); err != nil {
+		switch err {
+		case conf.ErrHelpWanted:
+			usage, err := conf.Usage("SALES", &cfg)
+			if err != nil {
+				return cfg, errors.Wrap(err, "generating config usage")
+			}
+			fmt.Println(usage)
+			return cfg, nil
+		case conf.ErrVersionWanted:
+			version, err := conf.VersionString("SALES", &cfg)
+			if err != nil {
+				return cfg, errors.Wrap(err, "generating config version")
+			}
+			fmt.Println(version)
+			return cfg, nil
+		}
+		return cfg, errors.Wrap(err, "parsing config")
+	}
+
+	out, err := conf.String(&cfg)
+	if err != nil {
+		return cfg, errors.Wrap(err, "generating config for output")
+	}
+	log.Infow("startup", "config", out)
+	return cfg, nil
+}
+
+func initExpvar(log *zap.SugaredLogger, cfg Config) {
+	expvar.NewString("build").Set(cfg.Version.SVN)
+	log.Infow("starting service", "version", cfg.Version.SVN)
+}
+
+func initAuth(log *zap.SugaredLogger, cfg Config) (*auth.Auth, error) {
+	log.Infow("startup", "status", "initializing authentication support")
+
+	// Construct a key store based on the key files stored in
+	// the specified directory.
+	ks, err := keystore.NewFS(os.DirFS(cfg.Auth.KeysFolder))
+	if err != nil {
+		return nil, errors.Wrap(err, "reading keys")
+	}
+
+	auth, err := auth.New(cfg.Auth.Algorithm, ks)
+	if err != nil {
+		return nil, errors.Wrap(err, "constructing auth")
+	}
+	return auth, nil
+}
+
+func initDatabase(log *zap.SugaredLogger, cfg Config) (*sqlx.DB, error) {
+	log.Infow("startup", "status", "initializing database support", "host", cfg.DB.Host)
+
+	db, err := database.Open(database.Config{
+		User:         cfg.DB.User,
+		Password:     cfg.DB.Password,
+		Host:         cfg.DB.Host,
+		Name:         cfg.DB.Name,
+		MaxIdleConns: cfg.DB.MaxIdleConns,
+		MaxOpenConns: cfg.DB.MaxOpenConns,
+		DisableTLS:   cfg.DB.DisableTLS,
+	})
+	if err != nil {
+		return nil, errors.Wrap(err, "connecting to db")
+	}
+	defer func() {
+		log.Infow("shutdown", "status", "stopping database support", "host", cfg.DB.Host)
+		db.Close()
+	}()
+
+	return db, nil
+}
+
+func initTracing(log *zap.SugaredLogger, cfg Config) error {
+	// WARNING: The current Init settings are using defaults which may not be
+	// compatible with your project. Please review the documentation for
+	// opentelemetry.
+
+	log.Infow("startup", "status", "initializing OT/Zipkin tracing support")
+
+	exporter, err := zipkin.New(
+		cfg.Zipkin.ReporterURI,
+		// zipkin.WithLogger(zap.NewStdLog(log)),
+	)
+	if err != nil {
+		return errors.Wrap(err, "creating new exporter")
+	}
+
+	traceProvider := trace.NewTracerProvider(
+		trace.WithSampler(trace.TraceIDRatioBased(cfg.Zipkin.Probability)),
+		trace.WithBatcher(exporter,
+			trace.WithMaxExportBatchSize(trace.DefaultMaxExportBatchSize),
+			trace.WithBatchTimeout(trace.DefaultBatchTimeout),
+			trace.WithMaxExportBatchSize(trace.DefaultMaxExportBatchSize),
+		),
+		trace.WithResource(
+			resource.NewWithAttributes(
+				semconv.SchemaURL,
+				semconv.ServiceNameKey.String(cfg.Zipkin.ServiceName),
+				attribute.String("exporter", "zipkin"),
+			),
+		),
+	)
+
+	// I can only get this working properly using the singleton :(
+	otel.SetTracerProvider(traceProvider)
+	defer traceProvider.Shutdown(context.Background())
+	return nil
+}
+
+func initDebugService(log *zap.SugaredLogger, cfg Config, db *sqlx.DB) {
+
+	log.Infow("startup", "status", "debug router started", "host", cfg.Web.DebugHost)
+
+	// The Debug function returns a mux to listen and serve on for all the debug
+	// related endpoints. This include the standard library endpoints.
+
+	// Construct the mux for the debug calls.
+	debugMux := handlers.DebugMux(cfg.Version.SVN, log, db)
+
+	// Start the service listening for debug requests.
+	// Not concerned with shutting this down with load shedding.
+	go func() {
+		if err := http.ListenAndServe(cfg.Web.DebugHost, debugMux); err != nil {
+			log.Errorw("shutdown", "status", "debug router closed", "host", cfg.Web.DebugHost, "ERROR", err)
+		}
+	}()
+}
+
+func initAPIService(log *zap.SugaredLogger, cfg Config, auth *auth.Auth, db *sqlx.DB) (http.Server, chan os.Signal, chan error) {
+	log.Infow("startup", "status", "initializing API support")
+
+	// Make a channel to listen for an interrupt or terminate signal from the OS.
+	// Use a buffered channel because the signal package requires it.
+	shutdown := make(chan os.Signal, 1)
+	signal.Notify(shutdown, syscall.SIGINT, syscall.SIGTERM)
+
+	// Construct the mux for the API calls.
+	apiMux := handlers.APIMux(handlers.APIMuxConfig{
+		Shutdown: shutdown,
+		Log:      log,
+		Metrics:  metrics.New(),
+		Auth:     auth,
+		DB:       db,
+	})
+
+	// Construct a server to service the requests against the mux.
+	api := http.Server{
+		Addr:         cfg.Web.APIHost,
+		Handler:      apiMux,
+		ReadTimeout:  cfg.Web.ReadTimeout,
+		WriteTimeout: cfg.Web.WriteTimeout,
+		IdleTimeout:  cfg.Web.IdleTimeout,
+		ErrorLog:     zap.NewStdLog(log.Desugar()),
+	}
+
+	// Make a channel to listen for errors coming from the listener. Use a
+	// buffered channel so the goroutine can exit if we don't collect this error.
+	serverErrors := make(chan error, 1)
+
+	// Start the service listening for api requests.
+	go func() {
+		log.Infow("startup", "status", "api router started", "host", api.Addr)
+		serverErrors <- api.ListenAndServe()
+	}()
+
+	return api, shutdown, serverErrors
+}
+
+func shut(log *zap.SugaredLogger, cfg Config, api http.Server, shutdown chan os.Signal, serverErrors chan error) error {
+	// Blocking main and waiting for shutdown.
+	select {
+	case err := <-serverErrors:
+		return errors.Wrap(err, "server error")
+
+	case sig := <-shutdown:
+		log.Infow("shutdown", "status", "shutdown started", "signal", sig)
+		defer log.Infow("shutdown", "status", "shutdown complete", "signal", sig)
+
+		// Give outstanding requests a deadline for completion.
+		ctx, cancel := context.WithTimeout(context.Background(), cfg.Web.ShutdownTimeout)
+		defer cancel()
+
+		// Asking listener to shutdown and shed load.
+		if err := api.Shutdown(ctx); err != nil {
+			api.Close()
+			return errors.Wrap(err, "could not stop server gracefully")
+		}
+	}
+	return nil
+}
