+++ business/config/models.go
@@ -1,0 +1,38 @@
+package config
+
+import (
+	"time"
+
+	"github.com/ardanlabs/conf"
+)
+
+type Config struct {
+	Args conf.Args
+	conf.Version
+	Web struct {
+		APIHost         string        `conf:"default:0.0.0.0:3000"`
+		DebugHost       string        `conf:"default:0.0.0.0:4000"`
+		ReadTimeout     time.Duration `conf:"default:5s"`
+		WriteTimeout    time.Duration `conf:"default:10s"`
+		IdleTimeout     time.Duration `conf:"default:120s"`
+		ShutdownTimeout time.Duration `conf:"default:20s"`
+	}
+	Auth struct {
+		KeysFolder string `conf:"default:../../zarf/keys/"`
+		Algorithm  string `conf:"default:RS256"`
+	}
+	DB struct {
+		User         string `conf:"default:postgres"`
+		Password     string `conf:"default:postgres,mask"`
+		Host         string `conf:"default:db"`
+		Name         string `conf:"default:postgres"`
+		MaxIdleConns int    `conf:"default:0"`
+		MaxOpenConns int    `conf:"default:0"`
+		DisableTLS   bool   `conf:"default:true"`
+	}
+	Zipkin struct {
+		ReporterURI string  `conf:"default:http://localhost:9411/api/v2/spans"`
+		ServiceName string  `conf:"default:sales-api"`
+		Probability float64 `conf:"default:0.05"`
+	}
+}
